iommufd: Add TSM_BIND
When TDISP-capable devices are passed through, the setup is first done
as usual but the device does not start doing encrypted DMA or accept
encrypted MMIO until its driver in a CoCo VM verifies the device and
only then enabled encrypted operations.

In order to verify the device, the TSM moves the device into a TDISP
CONFIG_LOCKED state which prohibits changes in the device's config space
so the driver can examine the device and move it to the RUN state if
the device is allowed into the TCB.

The TSM's operation for locking the device is called "BIND", add
an ioctl() to iommufd's vdevice to trigger the BIND operation.
The new ioctl() takes a made-up VM ID which KVM provides and
the TSM accepts.

iommufd_vdevice_tio_guest_request_ioctl

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>