resource: Mark encrypted MMIO resource on validation
Once TDISP MMIO validation succeeded, a resource needs to be mapped as
encrypted or a device rejects it.

Add encrypt_resource() which reserves the resource and marks it "validated"
(which is TDISP term meaning "must be accesses as encrypted"). The TSM module
is going to call this API to mark resources as validated.

Modify __ioremap_check_encrypted() to look for the new flag to allow
ioremap() correctly map encrypted resource.

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>
---

Should it also be IORESOURCE_BUSY?