KVM: SEV: Add TIO VMGEXIT
The SEV TIO spec defines a new TIO_GUEST_MESSAGE message to
provide a secure communication channel between a SNP VM and
the PSP, similar to the GHCB's GUEST_MESSAGE.

The defined messages provide way to read TDI info and do secure
MMIO/DMA setup.

On top of this, GHCB defines an extension to return certificates/
measurements/report and TDI run status to the VM.

The TIO_GUEST_MESSAGE handler also checks if a specific TDI bound
to the VM and exits the KVM to allow the userspace to bind it. This
KVM exit tells the user space an VM ID and guest device ID to be
bound, these two values do not have any special meaning (here they
are VMID=guest_ctxpage|asid and guest ID=guest device BDFn FIXME) and
are expected to be handled by the platform IOMMU driver which happens
in another patch.

Skip adjust_direct_map() in rmpupdate() for now as it fails on MMIO.

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>